+++ end-to-end/tests/testdata/s3/s3-collection.csv
@@ -1,2 +1,2 @@
 unique_id,node_id,title,title_language,alternative_title,member_of,contact_email,contact_name,collection_number,description,finding_aid
-,s3-collection-01,S3 Test Collection,English,,,,,,Collection for S3 tests;;eng,
+s3-collection-01,,S3 Test Collection,::name:English,,,,,,Collection for S3 tests;;eng,
